+++ packages/common/Img.svelte
@@ -5,6 +5,7 @@
   {alt}
   {...$$restProps}
 />
+<slot />
 
 <script>
   import { get_current_component } from 'svelte/internal';