+++ site/src/routes/demo/banner/_General.svelte
@@ -11,10 +11,6 @@
     <Checkbox bind:checked={mobileStacked} />
     <span slot="label">Mobile Stacked</span>
   </FormField>
-  <FormField>
-    <Checkbox bind:checked={icon} />
-    <span slot="label">Icon</span>
-  </FormField>
 </div>
 
 <pre class="status">Closed Reason: {closedReason}</pre>
@@ -27,36 +23,18 @@
       </Section>
     </Row>
   </TopAppBar>
-  {#if icon}
-    <Banner
-      bind:open
-      bind:centered
-      bind:mobileStacked
-      on:MDCBanner:closed={(event) => (closedReason = event.detail.reason)}
-    >
-      <Icon slot="icon" class="material-icons">favorite</Icon>
-      <Label slot="label">This is a banner with an icon and some actions.</Label
-      >
-      <svelte:fragment slot="actions">
-        <Button secondary>Secondary</Button>
-        <Button>Primary</Button>
-      </svelte:fragment>
-    </Banner>
-  {:else}
-    <Banner
-      bind:open
-      bind:centered
-      bind:mobileStacked
-      on:MDCBanner:closed={(event) => (closedReason = event.detail.reason)}
-    >
-      <Label slot="label">This is a banner with no icon and some actions.</Label
-      >
-      <svelte:fragment slot="actions">
-        <Button secondary>Secondary</Button>
-        <Button>Primary</Button>
-      </svelte:fragment>
-    </Banner>
-  {/if}
+  <Banner
+    bind:open
+    bind:centered
+    bind:mobileStacked
+    on:MDCBanner:closed={(event) => (closedReason = event.detail.reason)}
+  >
+    <Label slot="label">This is a banner with no icon and some actions.</Label>
+    <svelte:fragment slot="actions">
+      <Button secondary>Secondary</Button>
+      <Button>Primary</Button>
+    </svelte:fragment>
+  </Banner>
   <div>
     <img
       alt="Page content placeholder"
@@ -67,30 +45,15 @@
 </div>
 
 <script>
-  import Banner, { Label, Icon } from '@smui/banner';
+  import Banner, { Label } from '@smui/banner';
   import Button from '@smui/button';
   import TopAppBar, { Row, Section, Title } from '@smui/top-app-bar';
   import Checkbox from '@smui/checkbox';
   import FormField from '@smui/form-field';
 
-  let open = true;
+  let open = false;
   let centered = false;
   let mobileStacked = true;
-  let icon = true;
 
   let closedReason = 'None yet.';
 </script>
-
-<style>
-  .top-app-bar-container {
-    width: calc(100% - 24px);
-    height: 400px;
-    border: 1px solid
-      var(--mdc-theme-text-hint-on-background, rgba(0, 0, 0, 0.1));
-    margin: 0 18px 18px 0;
-    background-color: var(--mdc-theme-background, #fff);
-
-    overflow: auto;
-    display: inline-block;
-  }
-</style>