+++ site/src/routes/demo/textfield/index.svelte
@@ -62,6 +62,10 @@
     Both icons
   </Demo>
 
+  <Demo component={ConditionalIcons} file="textfield/_ConditionalIcons.svelte">
+    Conditional icons
+  </Demo>
+
   <Demo component={Textarea} file="textfield/_Textarea.svelte">Textarea</Demo>
 
   <Demo
@@ -103,6 +107,13 @@
   <Demo component={ManualSetup} file="textfield/_ManualSetup.svelte">
     Manual setup
   </Demo>
+
+  <Demo component={Solo} file="textfield/_Solo.svelte">
+    Solo style
+    <svelte:fragment slot="subtitle">
+      Using Paper and an Input outside a Textfield to create a "Solo" input
+    </svelte:fragment>
+  </Demo>
 </section>
 
 <script>
@@ -122,6 +133,7 @@
   import CharacterCount from './_CharacterCount.svelte';
   import HelperTextCharacterCount from './_HelperTextCharacterCount.svelte';
   import BothIcons from './_BothIcons.svelte';
+  import ConditionalIcons from './_ConditionalIcons.svelte';
   import Textarea from './_Textarea.svelte';
   import TextareaCharacterCount from './_TextareaCharacterCount.svelte';
   import FullWidth from './_FullWidth.svelte';
@@ -130,6 +142,7 @@
   import ElementsInLabel from './_ElementsInLabel.svelte';
   import DifferentTypes from './_DifferentTypes.svelte';
   import ManualSetup from './_ManualSetup.svelte';
+  import Solo from './_Solo.svelte';
 </script>
 
 <style>