+++ packages/common/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.2.0](https://github.com/hperrin/svelte-material-ui/compare/v4.1.0...v4.2.0) (2021-05-05)
+
+
+### Bug Fixes
+
+* events with modifiers called once for every event listener on event ([4234efa](https://github.com/hperrin/svelte-material-ui/commit/4234efa678ca9e33ab2c7647b9d4e634936309e0))
+
+
+### Features
+
+* allow customizing Label component, add Em and Strong elemental components ([5dd42e2](https://github.com/hperrin/svelte-material-ui/commit/5dd42e2829ef20328cfb5b6171191117286a8c05))
+
+
+
+
+
+# [4.1.0](https://github.com/hperrin/svelte-material-ui/compare/v4.0.0...v4.1.0) (2021-04-28)
+
+
+### Bug Fixes
+
+* warnings about unexpected slot in Img and Hr ([98a555f](https://github.com/hperrin/svelte-material-ui/commit/98a555f19273176099b8aa60b9e23028bb46e6e0))
+
+
+### Features
+
+* change event modifier syntax to allow use on all events ([1fe4f82](https://github.com/hperrin/svelte-material-ui/commit/1fe4f82ef29f88567a81a24a746ef7f82389e7dc))
+
+
+### Reverts
+
+* Revert "v4.1.0" ([54dfd48](https://github.com/hperrin/svelte-material-ui/commit/54dfd48fc0560843ccef9cb9821b12e251977c67))
+* Revert "v4.1.0" ([3e683d9](https://github.com/hperrin/svelte-material-ui/commit/3e683d9166e7df9eaab2756c9ace64c4dd547758))
+
+
+
+
+
 # [4.0.0](https://github.com/hperrin/svelte-material-ui/compare/v3.1.0...v4.0.0) (2021-04-24)